+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Added QFE in inHg and hPa to the ATIS report remarks
+
 ## [0.4.4] - 2018-11-21
 ### Fixed
 - Properly rotate and normalize the DCS reported wind direction